Box#encrypt: make nonce choosable
Sandom nonces are okay but unsuited for large numbers of messages with
the same key. In that case, make it a counter. If you can't guarantee to
provide a counter that will never use the same nonce again, you're
better off with XSalsa20 which uses a much larger nonce, so the
probability of randomly choosing the same nonce twice is much smaller.